+++ elastic/examples/async_sample/src/main.rs
@@ -1,12 +1,4 @@
-#[macro_use]
-extern crate serde_derive;
-#[macro_use]
-extern crate elastic_types_derive;
-#[macro_use]
-extern crate json_str;
-
 extern crate elastic_requests as req;
-extern crate elastic_responses as res;
 extern crate elastic_types;
 
 extern crate serde;
@@ -23,39 +15,19 @@ extern crate futures_cpupool;
 #[macro_use]
 extern crate quick_error;
 
-use std::collections::VecDeque;
-use std::path::Path;
-use std::str::{self, FromStr};
-use std::io::{Read, Error as IoError};
+use std::str;
+use std::io::Read;
 
 use req::BulkRequest;
 
 use tokio_core::reactor::Core;
-use tokio_proto::streaming::Body;
-use futures::{Future, Stream, Sink, Poll, Async};
-use futures::future::{ok, err, lazy, FutureResult};
-use futures::sync::mpsc::SendError;
+use futures::{Future, Stream};
+use futures::future::ok;
 use futures_cpupool::CpuPool;
 
-use memmap::{Mmap, MmapViewSync, Protection};
-use hyper::{Method, Error as HyperError};
-use hyper::header::ContentType;
-use hyper::client::Request;
-
-quick_error!{
-    #[derive(Debug)]
-    enum RequestError {
-        Io(err: IoError) {
-            from()
-        }
-        Body(err: SendError<ChunkResult>) {
-            from()
-        }
-        Request(err: HyperError) {
-            from()
-        }
-    }
-}
+mod body;
+mod error;
+mod hyper_req;
 
 fn main() {
     let url = "http://localhost:9200";
@@ -66,16 +38,16 @@ fn main() {
 
     // Create a `hyper` client expecting `MappedFileBody` bodies
     let client = hyper::client::Config::default()
-        .body::<MappedFileBody>()
+        .body::<body::MappedFileBody>()
         .build(&core.handle());
 
     // Get a future to buffer a bulk file
-    let (buffer_future, body) = mapped_file("./data/accounts.json").unwrap();
+    let (buffer_future, body) = body::mapped_file("./data/accounts.json").unwrap();
     let buffer_future = pool.spawn(buffer_future);
 
     // Get a future to send a bulk request
     let req = BulkRequest::new(body);
-    let req_future = client.request(hyper_req(&url, req))
+    let req_future = client.request(hyper_req::build(&url, req))
         .and_then(|res| {
             res.body()
                 .fold(Vec::new(), |mut buf, chunk| {
@@ -84,8 +56,7 @@ fn main() {
                         .map(|_| buf)
                 })
                 .and_then(|buf| {
-                    // TODO: Deserialize the response on the cpu pool
-
+                    // Write the response to `stdout`
                     println!("{:?}", str::from_utf8(&buf).unwrap());
 
                     ok(())
@@ -97,148 +68,4 @@ fn main() {
     let req_future = buffer_future.join(req_future);
 
     core.run(req_future).unwrap();
-}
-
-struct Chunk(MmapViewSync);
-
-impl AsRef<[u8]> for Chunk {
-    fn as_ref(&self) -> &[u8] {
-        unsafe { self.0.as_slice() }
-    }
-}
-
-type MappedFileBody = Body<Chunk, HyperError>;
-type ChunkResult = Result<Chunk, HyperError>;
-
-struct ChunkStream(VecDeque<ChunkResult>);
-
-impl Stream for ChunkStream {
-    type Item = ChunkResult;
-    type Error = SendError<ChunkResult>;
-
-    fn poll(&mut self) -> Poll<Option<Self::Item>, Self::Error> {
-        if let Some(slice) = self.0.pop_front() {
-            Ok(Async::Ready(Some(slice)))
-        } else {
-            Ok(Async::Ready(None))
-        }
-    }
-}
-
-/// A future for streaming the request body from a file.
-///
-/// The first item is a future that will stream chunks from the mapped file.
-/// The second item is the async body to use in the request.
-fn mapped_file<P>
-    (path: P)
-     -> Result<(Box<Future<Item = (), Error = RequestError> + Send>, MappedFileBody), IoError>
-    where P: AsRef<Path> + Send + 'static
-{
-    let (tx, rx) = MappedFileBody::pair();
-
-    let tx_future = lazy(move || {
-        let slices_future = map_file_to_chunks(path);
-
-        slices_future.and_then(|slices| {
-            let streamed = ChunkStream(slices);
-
-            tx.send_all(streamed).map_err(|e| e.into())
-        })
-    });
-
-    let tx_future = tx_future.and_then(|_| ok(()));
-
-    Ok((tx_future.boxed(), rx))
-}
-
-fn map_file_to_chunks<P>(path: P) -> FutureResult<VecDeque<ChunkResult>, RequestError>
-    where P: AsRef<Path>
-{
-    let file = match Mmap::open_path(path, Protection::Read) {
-        Ok(file) => file,
-        Err(e) => return err(e.into()),
-    };
-
-    let file = file.into_view_sync();
-
-    let total_len = file.len();
-
-    if total_len == 0 {
-        return ok(VecDeque::new())
-    }
-
-    let slice_size = 1024 * 16;
-
-    let num_slices = (total_len as f32 / slice_size as f32).ceil() as usize;
-
-    let mut slices = VecDeque::with_capacity(num_slices);
-
-    let mut next = Some(file);
-    while let Some(rest) = next {
-        next = match rest.len() {
-            // >1 chunk size left
-            len if len > slice_size => {
-                let (slice, rest) = match rest.split_at(slice_size) {
-                    Ok(split) => split,
-                    Err(e) => return err(e.into()),
-                };
-
-                slices.push_back(Ok(Chunk(slice)));
-
-                Some(rest)
-            },
-            // EOF
-            0 => {
-                None
-            },
-            // Last chunk
-            _ => {
-                slices.push_back(Ok(Chunk(rest)));
-
-                None
-            }
-        };
-    }
-
-    ok(slices)
-}
-
-/// Build a `hyper` request from an `elastic` request.
-fn hyper_req<I, B>(base_url: &str, req: I) -> Request<B>
-    where I: Into<req::HttpRequest<'static, B>>,
-          B: Stream<Error = hyper::Error> + 'static + ::std::fmt::Debug,
-          B::Item: AsRef<[u8]>
-{
-    let req = req.into();
-
-    let mut url = String::with_capacity(base_url.len() + req.url.len());
-
-    url.push_str(base_url);
-    url.push_str(&req.url);
-
-    let url = hyper::Uri::from_str(&url).unwrap();
-
-    let method = req.method;
-    let body = req.body;
-
-    let mut req = match method {
-        req::HttpMethod::Get => Request::<B>::new(Method::Get, url),
-        req::HttpMethod::Post => {
-            let mut req = Request::<B>::new(Method::Post, url);
-
-            if let Some(body) = body {
-                req.set_body(body);
-            }
-
-            req
-        }
-        _ => unimplemented!(),
-    };
-
-    {
-        let mut headers = req.headers_mut();
-        headers.set(ContentType::json())
-    }
-
-    req
 }
